requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
onnx>=1.14.0
onnxruntime-gpu>=1.15.0

# TensorRT and CUDA
nvidia-tensorrt>=8.6.0
pycuda>=2022.1
nvidia-ml-py3>=11.495.46

# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Additional utilities
psutil>=5.9.0
GPUtil>=1.4.0
requests>=2.31.0
aiofiles>=23.0.0
python-multipart>=0.0.6

# Configuration and serialization
pyyaml>=6.0
python-dotenv>=1.0.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0

# Data processing and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Enhanced performance
accelerate>=0.20.0
optimum>=1.9.0

# Optional: Additional ML utilities
scikit-learn>=1.3.0

# Optional: Database and caching support
redis>=4.6.0
sqlalchemy>=2.0.0
diskcache>=5.6.0

# Optional: Profiling and debugging
py-spy>=0.3.14
memory-profiler>=0.61.0

# Optional: ONNX optimization
onnxoptimizer>=0.3.0